requirements.txt
httpx
rich
duckduckgo-search
pyyaml
